resolved state
