+++ src/store/userSlice.js
@@ -22,35 +22,15 @@ export const followUser = createAsyncThunk(
   }
 );
 
-export const bookmarkPost = createAsyncThunk('posts/bookmarkPost', async({postId, encodedToken})=>{
-  try{
-      const response = await bookMarkService({postId, encodedToken})
-      return response.data.bookmarks
-  }catch(err){
-      console.error(err)
-  }
-})
-
-export const removeBookmark = createAsyncThunk('posts/removeBookmark', async({postId, encodedToken})=>{
-  try{
-    const response = await removeBookMarkService({postId, encodedToken})
-    return response.data.bookmarks
-  }catch(err){
-    console.error(err)
-}
-})
-
 export const userSlice = createSlice({
   name: "users",
   initialState: {
     users: [],
     currentUser: "",
-    currUserBookmark:[]
   },
   reducers: {
     updateCurrUser: (state, action) => {
       state.currentUser = action.payload;
-  
     },
   },
   extraReducers: {
@@ -74,25 +54,8 @@ export const userSlice = createSlice({
     [followUser.rejected]: (action) => {
       console.error(action);
     },
-    [bookmarkPost.fulfilled]:(state, action)=>{
-      state.currUserBookmark = [...state.users].filter((user)=>
-        action.payload[0].username === user.username
-      )
-      state.currUserBookmark[0].bookmarks = action.payload
+
   },
-  [bookmarkPost.rejected]:(action)=>{
-      console.log(action.payload)
-  }
-  },
-  [removeBookmark.fulfilled]:(state, action)=>{
-    state.currUserBookmark = [...state.users].filter((user)=>
-    action.payload[0].username === user.username
-  )
-  state.currUserBookmark[0].bookmarks = action.payload
-  },
-  [removeBookmark.rejected]:(action)=>{
-    console.log(action.payload)
-  }
 });
 
 export default userSlice.reducer;
